Refactor(#6) : 코드 아키텍처 전체 리펙토링
