+++ group_service/src/main/java/io/urdego/group_service/api/controller/GroupMemberController.java
@@ -1,4 +1,4 @@
-package io.urdego.group_service.api;
+package io.urdego.group_service.api.controller;
 
 public class GroupMemberController {
 }
