+++ group_service/src/main/java/io/urdego/group_service/api/service/GroupService.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class GroupMemberService {
+public class GroupService {
 }
