+++ group_service/src/main/java/io/urdego/group_service/common/exception/BaseException.java
@@ -1,0 +1,4 @@
+package io.urdego.group_service.common.exception;
+
+public class BaseException {
+}
