+++ group_service/src/main/java/io/urdego/group_service/common/exception/group/GroupException.java
@@ -1,0 +1,4 @@
+package io.urdego.group_service.common.exception.group;
+
+public class GroupException {
+}
