+++ group_service/src/main/java/io/urdego/group_service/common/exception/groupMember/GroupMemberException.java
@@ -1,0 +1,4 @@
+package io.urdego.group_service.common.exception.groupMember;
+
+public class GroupMemberException {
+}
