+++ group_service/src/main/java/io/urdego/group_service/domain/entity/groupMember/repository/GroupMemberRepository.java
@@ -1,8 +1,9 @@
-package io.urdego.group_service.domain.entity.group.repository;
+package io.urdego.group_service.domain.entity.groupMember.repository;
 
+import io.urdego.group_service.domain.entity.groupMember.GroupMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupMemberRepository extends JpaRepository<> {
+public interface GroupMemberRepository extends JpaRepository<GroupMember, Long> {
 }
